pipeline/requirements.txt
# MT Publishing Pipeline - Dependencies
# =====================================

# Core utilities
PyYAML>=6.0                    # Configuration file parsing
python-dateutil>=2.8.0         # Date handling
python-dotenv>=1.0.0           # Load environment variables from .env file

# EPUB Processing (Phase 1 & 4)
lxml>=4.9.0                    # XML/HTML parsing
beautifulsoup4>=4.12.0         # HTML parsing
Pillow>=10.0.0                 # Image processing
ebooklib>=0.18                 # EPUB manipulation
smartypants>=2.0.0             # Typographic quotes/dashes/ellipses (Phase 4 fail-safe)

# PDF Generation
reportlab>=4.0.0               # PDF creation

# Gemini API (Phase 2 - Translator & Phase 3 - Critics & Phase 5 - Narrator TTS)
google-genai>=0.5.0            # Google Gemini API SDK
google-generativeai>=0.8.0     # Google Generative AI SDK (for TTS)

# Retry logic and utilities
tenacity>=8.2.0                # Retry with exponential backoff
backoff>=2.2.0                 # Exponential backoff decorators

# Optional: Token counting
tiktoken>=0.5.0                # Token counting for cost estimation

# Vector Search (RAG Pattern Matching)
chromadb>=0.4.0                # Vector database for semantic search
google-cloud-aiplatform>=1.38.0  # Google AI embeddings (text-embedding-004)

# CLI Interface (TUI)
questionary>=2.0.0             # Interactive prompts with arrow-key navigation
rich>=13.0.0                   # Beautiful terminal output, progress bars, tables

# Development & Testing
pytest>=7.0.0                  # Testing framework
pytest-asyncio>=0.21.0         # Async test support
